+++ src/ChebyHash/ChebyHash64.hpp
@@ -9,5 +9,5 @@
 // authentication, key derivation, or secure storage.
 
 void round(std::uint64_t& x, std::uint64_t& y);
-std::uint32_t mix(std::uint64_t x, std::uint64_t y);
+std::uint64_t mix(std::uint64_t x, std::uint64_t y);
 void ChebyHashTester();
